Save eval result, not just formmated result
This is needed so that complicated objects like a graph
can be used by front-ends to render the object better
e.g. with say networkx-pyplot. And to use in retrieving
history when combining with future inputs.

Some black reformatting and other small changes have been made
as well. This syncronizes with the remove-front-ends branch.